feat(ui): support for axis left and right both
