+++ thirdeye-ui/src/app/components/visualizations/linear-axis-y/linear-axis-y.interfaces.ts
@@ -1,8 +1,10 @@
+import { Orientation } from "@visx/axis";
 import { ScaleLinear } from "d3-scale";
 
-export interface LinearAxisLeftProps {
+export interface LinearAxisYProps {
     top?: number;
     left?: number;
     numTicks?: number;
     scale: ScaleLinear<number, number>;
+    orientation?: Orientation;
 }
