+++ thirdeye-ui/src/app/components/visualizations/linear-axis-y/linear-axis-y.styles.ts
@@ -1,6 +1,6 @@
 import { makeStyles } from "@material-ui/core";
 
-export const useLinearAxisLeftStyles = makeStyles((theme) => ({
+export const useLinearAxisYStyles = makeStyles((theme) => ({
     tick: {
         ...theme.typography.overline,
     },
